Refactor: Fix Registration persistence, add DB Triggers, and standardize DTOs
Consolidated commit addressing critical persistence crashes during registration, enforcing data integrity via database triggers, and standardizing the API contract.

Backend & Persistence:

Implemented Persistable<UUID> on Account, UserProfile, Student, and Professor to correctly handle manual ID assignment.

Removed @MapsId and disabled CascadeType.PERSIST on Account relationships to prevent Hibernate from prematurely saving child entities with null IDs.

Updated AuthService to manually orchestrate the saving order (Account -> Profile -> Role) to ensure FK constraints are met.

Adjusted Entity mappings (insertable=false, nullable logic) to support both manual ID assignment and Database Trigger generation without conflict.

Database & Integrity:

Added trg_generate_professor_id to auto-generate IDs (e.g., 'P20250001') if the field is null.

Added trg_check_student_schedule_conflict to prevent students from enrolling in overlapping sections.

Refactored data.sql to be idempotent (safe to re-run without duplicate key errors).

DTOs, Mappers & API:

Enforced snake_case JSON keys using @JsonProperty annotations to resolve IDE warnings and ensure frontend compatibility.

Refactored EnrollmentMapper and SectionOfferingResponse to use a type-safe SectionStatus Enum (OPEN/FULL) instead of raw strings.

Updated AuthMapper to correctly handle default UUID generation and prevent null injection into non-nullable fields.

+++ src/main/kotlin/net/bscs22/schoolportal/services/SubjectService.kt
@@ -20,8 +20,8 @@ class SubjectService(
     fun createSubject(
         subjectCode: String,
         subjectName: String,
-        lecUnits: Long,
-        labUnits: Long
+        lecUnits: Int,
+        labUnits: Int
     ): Subject {
         if (subjectRepository.existsById(subjectCode)) {
             throw IllegalArgumentException("Subject code '$subjectCode' already exists.")
@@ -30,8 +30,8 @@ class SubjectService(
         val newSubject = Subject(
             subjectCode = subjectCode,
             subjectName = subjectName,
-            lecUnits = lecUnits,
-            labUnits = labUnits
+            lecUnits = lecUnits.toLong(), // Fix: Convert Int to Long
+            labUnits = labUnits.toLong()  // Fix: Convert Int to Long
         )
         return subjectRepository.save(newSubject)
     }
@@ -41,16 +41,16 @@ class SubjectService(
     fun updateSubject(
         code: String,
         subjectName: String,
-        lecUnits: Long,
-        labUnits: Long
+        lecUnits: Int,
+        labUnits: Int
     ): Subject {
         val subject = subjectRepository.findById(code)
             .orElseThrow { IllegalArgumentException("Subject '$code' not found.") }
 
         // Update fields
         subject.subjectName = subjectName
-        subject.lecUnits = lecUnits
-        subject.labUnits = labUnits
+        subject.lecUnits = lecUnits.toLong() // Fix: Convert Int to Long
+        subject.labUnits = labUnits.toLong() // Fix: Convert Int to Long
 
         return subjectRepository.save(subject)
     }
